Fix "+" so that it only means space in query components; disable test for S4T
The fix reveals a GS1 system issue with correct means to process
percent-encoded AI values when they appear in URIs.

It is not possible to unambiguously decode the example presented for S4T
without knowing how the decoded values from the URI components are to be
encoded within a particular AI's value, assuming that they can be, e.g.
where percent-encoding enabed for the AI.

Attempting to percent-encode all AI values containing characters outside
of CSET 82 would lead to ambiguity because in most AIs '%' is literal
whereas in certain AIs the '%' is an escape character. If
double-escaping is undesirable then a explicit indication is required
that covers the processing of data for each AI.

Such an indication would require a table of encoding methods to be
maintained, which is incompatible with the principles required for a
lightweight encoder that is suitable for embedding in offline devices
that have slow upgrade cycles.

In the absense of such an indication, the literal characters within an
AI value must always be URI encoded, and it is not possible to avoid
double-encoding, i.e.

- "+" in an AI value (percent encoded space) would be represented as the
three characters '%2B' in a URI (percent encoded "+"), with no
intermediate step of decoding the AI value "+" to space.
- "+" in a GS1 DL URI (percent encoded space) would be rejected since it
is now valid to put a space in an AI value.

Any scheme that avoids this needs to state what to do to get data in and
out of AI values in a meaningful for to the user for translation to /
from URIs (and other formats).